[Feature] 新增网络管理模块与窗口管理PID显示优化
1. 网络管理模块:
   - 实现了完整的网络连接列表展示功能，支持 TCP/UDP 协议、本地/远程 IP 及端口、连接状态及所属 PID。
   - 模块化设计：新增 ModuleNetworkManager.dll 客户端插件，采用 GetExtendedTcpTable 等 API 获取底层网络数据。
   - 主控端 UI：在主界面新增网络功能入口，添加 NetworkDialog 交互窗口及专用图标 (network.ico)。
   - 通讯协议：在 Config.h 中新增 CMD_NETWORK_LIST 指令，并完善 CommandHandler 的数据解析逻辑。
2. 窗口管理优化:
   - 在窗口列表中新增 PID 列，支持实时显示窗口关联的进程 ID。
   - 优化了客户端窗口枚举逻辑，确保 PID 数据采集的准确性。
3. 架构与工程优化:
   - 统一了各模块生成 DLL 的命名规范（ModuleXxxManager.dll）。
   - 修正了 MasterGUI.rc 中的资源引用路径及二进制文件打包逻辑。
   - 清理了 UI 代码中的冗余头文件引用，优化编译速度。
   - 调整 .gitignore 以排除超大文件 (avcodec-62.dll) 以适应 GitHub 限制。

+++ Master/UI/SettingsDialog.cpp
@@ -52,8 +52,8 @@ INT_PTR CALLBACK SettingsDlgProc(HWND hDlg, UINT message, WPARAM wParam, LPARAM 
             g_Settings.bEnableFrp = IsDlgButtonChecked(hDlg, IDC_RADIO_FRP_YES) == BST_CHECKED;
 
             // 保存桌面管理设置
-            g_Settings.screenCaptureMethod = SendDlgItemMessage(hDlg, IDC_COMBO_CAPTURE_METHOD, CB_GETCURSEL, 0, 0);
-            g_Settings.imageCompressMethod = SendDlgItemMessage(hDlg, IDC_COMBO_COMPRESS_METHOD, CB_GETCURSEL, 0, 0);
+            g_Settings.screenCaptureMethod = (int)SendDlgItemMessage(hDlg, IDC_COMBO_CAPTURE_METHOD, CB_GETCURSEL, 0, 0);
+            g_Settings.imageCompressMethod = (int)SendDlgItemMessage(hDlg, IDC_COMBO_COMPRESS_METHOD, CB_GETCURSEL, 0, 0);
             g_Settings.enableMultiMonitor = IsDlgButtonChecked(hDlg, IDC_RADIO_MULTIMON_YES) == BST_CHECKED;
             g_Settings.useDiffTransmission = IsDlgButtonChecked(hDlg, IDC_CHECK_DIFF_TRANS) == BST_CHECKED;
